Change telegram id field

+++ user/models.py
@@ -26,7 +26,7 @@ class User(AbstractUser):
     is_staff = BooleanField(default=False)
     # is_agent = BooleanField(default=False)
     is_client = BooleanField(default=False)
-    telegram_id = models.IntegerField(null=True, blank=True)
+    telegram_id = models.BigIntegerField(null=True, blank=True)
     referal_code = models.CharField(max_length=255, null=True, blank=True)
     refered = models.ForeignKey(
         "self", on_delete=models.SET_NULL, null=True, blank=True
